drm/omap: hdmi4: fix use of uninitialized var
commit df91637f201cd963bfcec4b389e5a1693dd6aa8c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

If use_mclk is false, mclk_mode is written to a register without
initialization. This doesn't cause any ill effects as the written value
is not used when use_mclk is false.

To fix this, write use_mclk only when use_mclk is true.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>