watchdog: Add K3 RTI watchdog support
commit bde7fba118461f46cb2e8de754db0847fa19d807 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Texas Instruments K3 SoCs contain an RTI (Real Time Interrupt) module
which can be used as a watchdog. This IP provides a support for
windowed watchdog mode, in which the watchdog must be petted within
a certain time window. If it is petted either too soon, or too late,
a watchdog error will be triggered.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>