ARM: dts: omap3: fix DPLL4 M4 divider max value
commit f5869190667951720f8c1ec4638bff4c682a3a4e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The maximum divider value for DPLL4 M4 divider appears wrong. For most
OMAP3 family SoCs this is 16, but it is defined as 32, which is maybe
only valid for omap36xx. To avoid any overflows in trying to write this
register, set the max to 16 for all omap3 family, except omap36xx. For
omap36xx the maximum is set to 31, as it appears value 32 is not working
properly.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Tested-by: Adam Ford <aford173@gmail.com>
Acked-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>