bus: ti-sysc: re-order reset and main clock controls
commit bf59ebbeac1fdd7d398d5f808f2cf239c8e61947 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The main clocks and reset controls have a hardware level dependency,
where one can't transition state without the other one transitioning.
Because we don't have the dependency implemented in software, we must
ensure the ordering of these two is done properly; they way this is
handled is that clocks transition on software level without delay,
and the status is only polled on reset side. Because of this, we must
re-order the main clock and reset handling on the ti-sysc driver.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>