ARM: dts: dra7: add vpe clkctrl node
commit 6c494153df36566b0a3d66038e63dd7b739d9ec3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add clkctrl nodes for VPE module.

Note that because of the current dts node name dependency for mapping to
clock domain, we must still use "vpe-clkctrl@" naming instead of generic
"clock@" naming for the node. And because of this, it's probably best to
apply the dts node addition together along with the other clock changes.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Acked-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>