crypto: omap-sham: fix split update cases with cryptomgr tests
commit db96de2fae5c6190e1d423dbdff377563e9358d4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The updated crypto manager finds a couple of new bugs from the omap-sham
driver. Basically the split update cases fail to calculate the amount of
data to be sent properly, leading into failed results and hangs with the
hw accelerator.

To fix these, the buffer handling needs to be fixed, but we do some cleanup
for the code at the same time to cut away some unnecessary code so that
it is easier to fix.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>