media: ti-vpe: cal: Add per platform data support
commit 55596690371adc676758df3dbf1d34a05399630f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

First this patch adds a method to access the CTRL_CORE_CAMERRX_CONTROL
register to use the syscon mechanism. For backward compatibility we also
handle using the existing camerrx_control "reg" entry if a syscon node
is not found.

In addition the register bit layout for the CTRL_CORE_CAMERRX_CONTROL
changes depending on the device. In order to support this we need to use
a register access scheme based on data configuration instead of using
static macro.

In this case we make use of the regmap facility and create data set
based on the various device and phy available.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>