media: ti-vpe: csc: fix single vs multiplanar format handling
commit 1d562382cba411d70607102e8f5e359e39acb1e0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In csc_set_coeff() we were previously checking if the format type was
V4L2_BUF_TYPE_VIDEO_OUTPUT or V4L2_BUF_TYPE_VIDEO_OUTPUT_MPLANE as a
mean to figure out if the format was multiplanar or not. This was not
entirely accurate as the format type could also be a CAPTURE type
instead.

Fix this by removing the 'switch' construct and replacing them with
"if (V4L2_TYPE_IS_MULTIPLANAR(type))" instead.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>