dmaengine: ti: edma: Add support for handling reserved channels
commit 9a56a919b91990a410980fcc98f53f0e8855d6ad from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 31f4b28f6c41f734957ea66ac84c6abb69e696f0 upstream.

Like paRAM slots, channels could be used by other cores and in this case
we need to make sure that the driver do not alter these channels.

Handle the generic dma-channel-mask property to mark channels in a bitmap
which can not be used by Linux and convert the legacy rsv_chans if it is
provided by platform_data.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Link: https://lore.kernel.org/r/20191025073056.25450-4-peter.ujfalusi@ti.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>