arm64: dts: ti: k3-j721e-common-proc-board: Add mailboxes to C71x DSP
commit 814a1eb789e1bc36391639c86a7c3e791d3917c5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the required 'mboxes' property to the C71x DSP processor on the TI
J721E common processor board. The mailboxes and some shared memory are
required for running the Remote Processor Messaging (RPMsg) stack between
the host processor and each of the R5Fs. The chosen sub-mailboxes match
the values used in the current firmware images. This can be changed,
if needed, as per the system integration needs after making appropriate
changes on the firmware side as well.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>