ARM: OMAP2+: sleep43xx: Call secure suspend/resume handlers
commit e5c54e55792949a8f299a1c261b25cfe204e5cd2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 8ab871f8bdfcca627802b6992e556fba5722a268 upstream.

During suspend CPU context may be lost in both non-secure and secure CPU
states. The kernel can handle saving and restoring the non-secure context
but must call into the secure side to allow it to save any context it may
lose. Add these calls here.

Note that on systems with OP-TEE available the suspend call is issued to
OP-TEE using the ARM SMCCC, but the resume call is always issued to the
ROM. This is because on waking from suspend the ROM is restored as the
secure monitor. It is this resume call that instructs the ROM to restore
OP-TEE, all subsequent calls will be handled by OP-TEE and should use the
ARM SMCCC.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Acked-by: Dave Gerlach <d-gerlach@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>