phy: cadence: Sierra: Configure both lane cdb and common cdb registers for external SSC
commit 46567ec37a561319c719ee79263c23e58afb2467 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 871002d788817eb4cd0cd03101d284c3db06ed74 upstream.

The existing configuration done in Cadence Sierra driver is only for
reference and is not used in any platforms. Remove them and configure
both lane cdb and common cdb registers to be used with external
SSC configuration. This is validated in TI J721E platform.

Signed-off-by: Anil Varughese <aniljoy@cadence.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>