ARM: dts: keystone-k2g: Add the PRU-ICSS nodes
commit e75cded5c8c4e08a1bb75779a93d6278f9887fc9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the DT nodes for the PRU-ICSS0 and PRU-ICSS1 processor subsystems
that are present on the 66AK2G SoC. The two PRU-ICSSs are identical
to each other. Each PRU-ICSS instance is represented by a pruss node
and other child nodes. These nodes are enabled by default.

The PRU-ICSSs on 66AK2G are very similar to the PRUSS on the AM57xx
SoCs except for larger Shared Data RAM and the lack of a PRU-ICSS
crossbar. There are also few other minor integration differences
w.r.t IPC mechansims that can be attributed to the architecture
differences between Keystone and OMAP families.

The PRUSS subsystem node contains the entire address space. The
various sub-modules of the PRU-ICSS are represented as individual
child nodes (so platform devices themselves) of the PRUSS subsystem
node. These include the two PRU cores and the interrupt controller.
The Industrial Ethernet Peripheral (IEP), the Real Time Media
Independent Interface controller (MII_RT), and the CFG sub-module
are represented as syscon nodes. All the Data RAMs are represented
within a child node of its own named 'memories' without any compatible.

The DT nodes use all standard properties. The regs property in the
PRU nodes define the addresses for the Instruction RAM, the Debug
and Control sub-modules for that PRU core. The firmware for each
PRU core is defined through a 'firmware-name' property.

The default names for the firmware images for each PRU core are
defined as follows (these can be adjusted either in derivative
board dts files or through sysfs at runtime if required):
    PRU-ICSS0 PRU0 Core: k2g-pru0_0-fw
    PRU-ICSS0 PRU1 Core: k2g-pru0_1-fw
    PRU-ICSS1 PRU0 Core: k2g-pru1_0-fw
    PRU-ICSS1 PRU1 Core: k2g-pru1_1-fw

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>