mfd: syscon: Use a unique name with regmap_config
commit 517e98f3dcb09826e9aef1d9033ed494bccbe159 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The DT node full name is currently being used in regmap_config
which in turn is used to create the regmap debugfs directories.
This name however is not guaranteed to be unique and the regmap
debugfs registration can fail in the cases where the syscon nodes
have the same unit-address but are present in different DT node
hierarchies. Replace this logic using the syscon reg resource
address instead (inspired from logic used while creating platform
devices) to ensure a unique name is given for each syscon.

[s-anna@ti.com: cherry-pick https://patchwork.kernel.org/patch/11353353/ ]
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>