ARM: dts: am437x-idk: Enable PRU-ICSS nodes
commit 524f7f97a94165a35b6a229194ad8103ef1f9413 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The AM437x IDK EVM board uses a AM437x SoC that supports two PRU-ICSS
instances. The PRU-ICSS target module node was left in disabled state
in the base am4372.dtsi file. Enable the PRU-ICSS target module node
to support both the PRU-ICSS1 and PRU-ICSS0 instances on this board.
The PRUSS nodes and most of their child nodes are already enabled in
the base dtsi file, and so become effective automatically with the
enabling of this PRU-ICSS target module node.

The corresponding PRU nodes can be disabled later on if there are
no use-cases defined to use a particular PRU core or the whole
PRU-ICSS subsystem itself if both its PRU cores are unused.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>