dma-buf: Move locks to only protect framework data
commit 758b4543bd776f1f95b740fab433fe5418caf9f7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The DMA-BUF framework mutex lock is used to protect the attachments list,
but in some spots it also surrounds the exporter callback attach. This is
not needed and causes a lockdep warning when exporters call back into
the DMA-BUF framework. Move the locking to only guard the framework
data it is meant to protect.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>