dma-direct: provide mmap and get_sgtable method overrides
commit 9b052468ad467c7e789a78934bd01f54c4a496b3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 34dc0ea6bc960f1f57b2148f01a3f4da23f87013 upstream.

For dma-direct we know that the DMA address is an encoding of the
physical address that we can trivially decode.  Use that fact to
provide implementations that do not need the arch_dma_coherent_to_pfn
architecture hook.  Note that we still can only support mmap of
non-coherent memory only if the architecture provides a way to set an
uncached bit in the page tables.  This must be true for architectures
that use the generic remap helpers, but other architectures can also
manually select it.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Max Filippov <jcmvbkbc@gmail.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>