ARM: dts: beagle-x15-common: Enable PRU-ICSS nodes
commit a2b463a2e348ad7b3ae6c21f97ca90d700ef3efd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The two PRU-ICSS processor subsystem interconnect target module nodes
were left in disabled state in the dra7-pruss.dtsi file. The PRU-ICSSs
are supported on only the AM57xx SoCs, so enable these nodes to support
both the PRU-ICSS1 and PRU-ICSS2 instances on all the BeagleBoard-X15
and AM57xx GP EVM boards. The PRUSS sub-system and PRU nodes are already
enabled in the dtsi file, and so become effective automatically with
the enabling of these target module interconnect nodes.

The corresponding PRU nodes can be disabled later on if there are
no use-cases defined to use a particular PRU core or the whole
PRU-ICSS subsystem itself if both its PRU cores are unused.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>