ARM: dts: AM4372: Add the PRU-ICSS1 DT node
commit fbb00a71f866eb07fc4e7f0b7a3f73c127065b74 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the DT node for the PRU-ICSS1 instance on the AM437x family
of SoCs. Each PRU-ICSS instance is represented by a pruss node
and other child nodes. The PRU-ICSS instances are not supported
on AM4372 SoC though in the AM437x family, so the nodes are added
under the corresponding disabled interconnect target module node
in the common am4372 dtsi file. The target module node should be
enabled in only those derivative board files that use a SoC
containing the PRU-ICSS.

The PRU-ICSS1 on AM437x is very similar to the PRUSS in AM33xx,
except for variations in the RAM sizes, bus addresses and the
number of interrupts coming into the MPU INTC (host interrupt
7 is routed to the other PRUSS instead of MPU).

The PRUSS subsystem node contains the entire address space. The
various sub-modules of the PRU-ICSS are represented as individual
child nodes (so platform devices themselves) of the PRUSS subsystem
node. These include the two PRU cores and the interrupt controller.
The Industrial Ethernet Peripheral (IEP), the Real Time Media
Independent Interface controller (MII_RT), and the CFG sub-module
are represented as syscon nodes. All the Data RAMs are represented
within a child node of its own named 'memories' without any compatible.

The DT nodes use all standard properties. The regs property in the
PRU nodes define the addresses for the Instruction RAM, the Debug
and Control sub-modules for that PRU core. The firmware for each
PRU core is defined through a 'firmware-name' property.

The default names for the firmware images for each PRU core are
defined as follows (these can be adjusted either in derivative
board dts files or through sysfs at runtime if required):
     PRU-ICSS1 PRU0 Core: am437x-pru1_0-fw
     PRU-ICSS1 PRU1 Core: am437x-pru1_1-fw

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>