HACK: dt-bindings: misc: Add ti,dma_buf_phys binding doc
commit d1d3663e1838d9555c51907730d5e639cbe1eecc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Driver allowing user-space attaching of DMA-BUFs returning CPU physical
addresses. The reasoning for making this act like a regular device
described by DT is so the virtual device that binds the buffer can be
made to act as if it is out on a bus or behind an IOMMU, for example.

Add this binding here.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>