HACK: misc: Add dma-buf to physical address exporter
commit 3229caddd169cfb60c53f2ecfd0ec5e6d6226d19 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This is driver allows user-space to attach a DMA-BUF and
receive back its CPU physical address. This is a temporary
solution to allow CMEM like functionality from allocated
buffers. This is a hack and this will be removed when proper
solutions are implemented.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>