arm64: dts: ti: k3-j721e: Enable Super-Speed support for USB0
commit d53146bdb941106d60b422ee13d04937ad98197e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

USB0 supports super-speed mode on the EVM. Enable that.
On the EVM, USB0 uses SERDES3 for super-speed lane.

Since USB0 is a type-C port, it needs to support lane swapping
for cable flip support. This is provided using SERDES lane
swap feature. Provide the Type-C cable orientation GPIO
to the SERDES Wrapper driver.

The Type-C compainon chip on the board needs ~133ms (tCCB_DEFAULT)
to debounce the CC lines in order to detect attach and plug
orientation and reflect the correct DIR status. [1]

We need to wait for 700ms before sampling the Type-C DIR line as
there seems to be more delay on the actual boards.

[1] http://www.ti.com/lit/ds/symlink/tusb321.pdf
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>