ARM: OMAP2+: am43xx: Add lcdc clockdomain
commit 5399f7e84eeb328d1ccb79e9c309e0bfe3c3d30d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

As described in AM437x TRM, spruhl7h, Revised January 2018, there is
an LCDC clockdomain present in the PER power domain. Although it is
entirely unused on AM437x, it should be defined along with the other
clockdomains so it can be shut off by Linux as there are no users.

Reported-by: Munan Xu <munan@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>